chore: replace array of users with hashmap of users
This should be much faster for retrieving a user. I am very proud of myself here.
The idea was, when adding an exercise, you are provided an _id of a user.
Why would I use array.find(_id) or other loops to go through the whole array,
when I can just use hashmap in the first place? Yay.