feat: Validate date, return valid object for exercise
Validating the date turned out to be a huge pain in the ass.
I ended up using express-validator, which offers a simple way to check if a
form field post input is empty. finally, the date validation worked, and the
res.json that i returned was finally accepted by the freecodecamp validation.

I also abandoned the hash map because I thought it might be the cause of the problem.
That didnt change anything, but I am so frustrated that I will currently not go back to the hash map.
Maybe when i finished this.
I'm proud of myself because I found a way out of this fucking hell on my own.